// Application/Frigorino.Web/ClientApp/src/lib/api/models/MemberResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { HouseholdRole } from './HouseholdRole';
export type MemberResponse = {
    externalId: string;
    name: string;
    email: string;
    role: HouseholdRole;
    joinedAt: string;
};